+++ udata/core/user/api_fields.py
@@ -91,6 +91,16 @@ apikey_fields = api.model('ApiKey', {
 
 user_page_fields = api.model('UserPage', fields.pager(user_fields))
 
+user_suggestion_fields = api.model('UserSuggestion', {
+    'id': fields.String(description='The user identifier', readonly=True),
+    'first_name': fields.String(description='The user first name',
+                                readonly=True),
+    'last_name': fields.String(description='The user last name',
+                               readonly=True),
+    'avatar_url': fields.String(description='The user avatar URL'),
+    'slug': fields.String(
+        description='The user permalink string', readonly=True),
+})
 
 notifications_fields = api.model('Notification', {
     'type': fields.String(description='The notification type', readonly=True),